+++ keyboards/kyria/keymaps/fi/snow.h
@@ -22,8 +22,8 @@
 #define VARIANCE 10
 
 typedef struct {
-    int32_t x;
-    int32_t y;
+    int16_t x;
+    int16_t y;
 } ivec2_t;
 
 typedef struct {
